+++ TaleWorlds.CampaignSystem/TaleWorlds.CampaignSystem.ComponentInterfaces/PrisonerRecruitmentCalculationModel.cs
@@ -9,7 +9,7 @@ public abstract class PrisonerRecruitmentCalculationModel : MBGameModel<Prisoner
 
 	public abstract ExplainedNumber GetConformityChangePerHour(PartyBase party, CharacterObject character);
 
-	public abstract int GetPrisonerRecruitmentMoraleEffect(PartyBase party, CharacterObject character, int num);
+	public abstract float GetPrisonerRecruitmentMoraleEffect(PartyBase party, CharacterObject character, int num);
 
 	public abstract bool IsPrisonerRecruitable(PartyBase party, CharacterObject character, out int conformityNeeded);
